+++ src/hdk/index.ts
@@ -1,4 +1,5 @@
+export * from "./action.js";
 export * from "./capabilities.js";
 export * from "./dht-ops.js";
 export * from "./entry.js";
-export * from "./header.js";
+export * from "./record.js";