+++ src/hdk/record.ts
@@ -1,0 +1,22 @@
+import { SignedActionHashed } from "./action.js";
+
+export type Record = {
+  signed_action: SignedActionHashed;
+  entry: RecordEntry;
+};
+
+export type RecordEntry =
+  | {
+      Present: {
+        entry: Uint8Array;
+      };
+    }
+  | {
+      Hidden: void;
+    }
+  | {
+      NotApplicable: void;
+    }
+  | {
+      NotStored: void;
+    };